pyproject.toml
[tool.poetry]
name = "full-stack-flask"
version = "0.1.0"
description = ""
authors = ["RedLotus <ssfdust@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
marshmallow = "*"
flask-smorest = "*"
flask-sqlalchemy = "^2.4"
flask-marshmallow = "^0.10.1"
flask-admin = { git = "https://github.com/flask-admin/flask-admin.git", branch = "bootstrap4" }
sqlalchemy-utils = "^0.34.1"
arrow = "^0.14.4"
loguru = "^0.3.2"
marshmallow-sqlalchemy = "^0.17.0"
flask-migrate = "^2.5"
flask-jwt-extended = "^3.21"
flask-security-too = "^3.2"
bcrypt = "^3.1"
flask-babel = "^0.12.2"
psycopg2-binary = "^2.8"
flask-mail = "^0.9.1"
celery = "^4.3"
flask-socketio = "^4.2"
ujson = "^1.35"
toml = "^0.10.0"
eventlet = "^0.25.1"
captcha = "^0.3.0"
flask-mongoengine = "^0.9.5"
invoke = "^1.3"

[tool.poetry.dev-dependencies]
ipython = "^7.8.0"
pytest = "^5.2.1"
pytest-sugar = "^0.9.2"
colorlog = "^4.0"
python-dotenv = "^0.10.3"
supervisor = "^4.0"
pytest-cov = "^2.8.1"
pdbpp = "^0.10.2"
pytest-lazy-fixture = "^0.6.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
